+++ CHANGES.rst
@@ -8,6 +8,8 @@ Change log
 - Use a default configuration file named ``toll.ini``. This can be overwritten
   using ``-c`` when calling `toll`.
 
+- Allow to specify multiple command which should be called. (See ``--help``.)
+
 - Make package compatible with Python 2.7, 3.3, 3.4 and PyPy.